requirements.txt
Fabric==1.8.0
Jinja2==2.7.1
Markdown==2.3.1
Pillow==2.2.1
PyYAML==3.10
